requirements.txt
fastapi
uvicorn
python-multipart
Pillow
numpy
torch
clip-anytorch
azure-search-documents